pyproject.toml
[project]
name = "q-agents"
version = "0.1.0"
description = "Multi-agent financial analysis system built with LangGraph."
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "mcp[cli]",
    "langchain",
    "langgraph",
    "langchain-openai",
    "chromadb",
    "langchain-chroma",
    "pandas",
    "pandas-market-calendars>=4.3.3",
    "yfinance",
    "tavily-python",
    "ddgs",
    "python-dotenv",
    "pytest",
    "playwright",
    "numpy>=2.2.6",
    "pandas-ta>=0.4.71b0",
    "lightgbm>=4.6.0",
    "shap>=0.51.0",
    "scikit-learn>=1.8.0",
    "langchain-community>=0.4.1",
    "anthropic>=0.85.0",
    "langchain-anthropic>=1.4.0",
    "langchain-huggingface>=1.2.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "httpx>=0.28.1",
    "apscheduler>=3.10.4",
    "python-okx>=0.4.1",
    "tenacity>=9.1.4",
    "redis>=5.0.0",
    "arq>=0.26.0",
    "hiredis>=2.3.0",
    "msgpack>=1.0.0",
    "pyarrow>=15.0.0",
    "ruff>=0.15.7",
]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.uv]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "B"]
ignore = [
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["E402"]
"tests/*.py" = ["F841", "E402"]
"app/ml/*.py" = ["N803", "N806"]
"mcp_servers/**/*.py" = ["E402"]